add testing data and implement utils

+++ src/main/java/com/agoda/compress/Compressable.java
@@ -4,7 +4,7 @@ import java.io.IOException;
 
 public interface Compressable {
 
-	void compress(String inputDir, String outputDir, Integer fileSize) throws IOException;
+	void compress(String inputDir, String outputDir, long fileSize) throws IOException;
 
 	void extract(String inputDir, String outputDir);
 }
